api [nfc]: Move UserSettingName and Emojiset to model, from initial_snapshot
These are also used in events.
